courssM/schedules/forms.py
from django import forms
from .models import *

class SchedulesForm(forms.ModelForm):
    date = forms.DateField(widget=forms.DateInput(attrs={'type': 'date'}))
    time = forms.TimeField(widget=forms.TimeInput(attrs={'type': 'time'}))
    class Meta:
        model = Schedules
        fields =  ['date','time','description']